fix: seed production
